% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_and_leaf_funs.R
\name{getLayout}
\alias{getLayout}
\title{Extract the layout from an igraph object.}
\usage{
getLayout(g)
}
\arguments{
\item{g}{An igraph object.}
}
\value{
The layout from the igraph object.
}
\description{
Extract the layout from an igraph object.
}
\examples{
getLayout(gg)
}
