% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rings.R, R/treeAndLeaf.R
\name{treeAndLeaf}
\alias{treeAndLeaf}
\title{TreeAndLeaf Layout}
\usage{
treeAndLeaf(gg, size = "small")

treeAndLeaf(gg, size = "small")
}
\arguments{
\item{gg}{An igraph object.}

\item{zoom}{An integer.}
}
\value{
The layout in the form of a matrix.
}
\description{
\code{rings}
}
\examples{
treeAndLeaf(gg)
}
