% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayout.R
\name{getLayout}
\alias{getLayout}
\title{Extracts the layout from an igraph object}
\usage{
getLayout(g)
}
\arguments{
\item{g}{an igraph object}
}
\value{
A matrix of positions
}
\description{
Function for extracting the layout from an igraph object for posterior usage
}
\examples{
layout <- getLayout(g)

}
\seealso{
\code{\link[RedeR:addGraph]{addGraph}},
\code{\link[igraph]{igraph}}
}
