% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeAndLeaf.R
\name{treeAndLeaf}
\alias{treeAndLeaf}
\title{Initial layout creation for the TreeAndLeaf}
\usage{
treeAndLeaf(gg, size = "small")
}
\arguments{
\item{gg}{an hclust or igraph object, containing the dendrogram to
be reorganized}

\item{size}{a string that describes the size of the dendrogram. "small" = less
than 100 nodes; "medium" = less than 500 nodes; "large" = 500+ nodes.
This is just a reference point, use the size that gives the best results.}
}
\value{
A matrix of the nodes positions
}
\description{
Returns a matrix of positions to be relaxed by the force based algorithm
implemented in the RedeR package
}
\examples{
hc <- hclust(dist(USArrests), "ave")
layout <- treeAndLeaf(hc)

}
\seealso{
\code{\link[RedeR:addGraph]{addGraph}}
}
